Test current serialization procedure
In manual ad-hoc tests a flaw was found.
This test explicitly tests for the buggy behaviour.

When a computed MAC contains the same character as the chosen separator,
the number of parts received after splitting the token is more than were
used to make the token in the first place.